Logic for disabled balances with null referred_by value

+++ internal/service/referralid/main_test.go
@@ -6,7 +6,7 @@ func TestNew(t *testing.T) {
 	tests := []struct {
 		name  string
 		did   string
-		index uint
+		index uint64
 		want  string
 	}{
 		{
